Code clean.

+++ MuseEngine/Muse/src/Core/Renderer/Buffer/BufferLayout.cpp
@@ -5,14 +5,14 @@ namespace Muse
 {
     void BufferLayout::CalculateOffsetAndStride()
     {
-        uint32_t offset = 0;
-        m_Stride = 0;
+		size_t offset = 0;
+		m_Stride = 0;
 
-        for (auto& element : m_Elements)
-        {
-            element.Offset = offset;
-            offset += element.Size;
-            m_Stride += element.Size;
-        }
+		for (auto& element : m_Elements)
+		{
+			element.Offset = offset;
+			offset += element.Size;
+			m_Stride += element.Size;
+		}
     }
 }
